Refactor `dot` and `ls` code.
Code common to Stack's `dot` and `ls` commands is moved to separate modules. `DotOpts` is moved to `Stack.Types.DotOpts`. `DependencyTree`, `DotPayload`, `licenseText` and `versionText` are moved to `Stack.Types.DependencyTree`. `createPrunedDependencyGraph` (and associated helper functions) is moved to `Stack.DependencyGraph`.

`Stack.Dot.DotException` is renamed `Stack.DependencyGraph.DependencyGraphException`.

`ls`-related parsers are moved from `Stack.Options.DotParser` to `Stack.Options.LsParser`. Related types are moved from `Stack.Dot` to `Stack.Ls`.

`dot` is renamed `dotCmd` to be consistent with the naming of similar functions.
